Generalize the activity for Donation and Request search
