+++ app/src/main/res/layout/activity_donation_search.xml
@@ -32,6 +32,7 @@
         android:layout_marginTop="195dp"
         android:layout_marginEnd="3dp"
         android:layout_marginBottom="67dp"
+        android:fillViewport="true"
         app:layout_constraintBottom_toTopOf="@+id/donate"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -39,8 +40,13 @@
 
         <LinearLayout
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical" />
+            android:layout_height="match_parent">
+
+            <ListView
+                android:id="@+id/donationListings"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent" />
+        </LinearLayout>
     </ScrollView>
 
     <Button
