+++ app/src/main/res/layout/activity_request_search.xml
@@ -32,17 +32,21 @@
         android:layout_marginTop="32dp"
         android:layout_marginEnd="3dp"
         android:layout_marginBottom="69dp"
+        android:fillViewport="true"
         app:layout_constraintBottom_toTopOf="@+id/back"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textView">
-        <ListView
-            android:id="@+id/helpListingList"
+        <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            />
+            android:orientation="vertical">
+            <ListView
+                android:id="@+id/requestListings"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"/>
+        </LinearLayout>
     </ScrollView>
-
     <Button
         android:id="@+id/back"
         android:layout_width="wrap_content"
